requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
supabase==1.0.3
httpx==0.23.3
python-dotenv==1.0.0
